pyproject.toml
[project]
name = "scp-stage1"
version = "0.1.0"
description = "Stage 1: Economic domain continued pretraining (Qwen3.5-4B + LoRA)"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.5.0",
    "transformers>=4.57.0",
    "peft>=0.15.0",
    "datasets>=3.0.0",
    "accelerate>=1.3.0",
    "trl>=0.15.0",
    "unsloth",
    "wandb",
    "sentencepiece",
    "tiktoken",
    "safetensors",
    "pyyaml",
    "tqdm",
    "numpy",
]

[project.optional-dependencies]
flash = ["flash-attn>=2.7.0", "triton>=3.0.0"]

[tool.setuptools.packages.find]
include = ["src*"]
